# Conflicts:
#	src/item/Column.java
#	src/item/Item.java

+++ src/review/Review.java
@@ -23,7 +23,7 @@ public class Review {
     }
 
     //complexitat O( 1 )
-    public void setPoints(int points){
+    public void setPoints(double points){
         this.points = points;
     }
